+++ apps/games-list-screen/vite.config.ts
@@ -1,11 +1,11 @@
-import { defineConfig, type PluginOption } from 'vite'
-import react from '@vitejs/plugin-react'
+import { defineConfig, type PluginOption } from 'vite';
+import react from '@vitejs/plugin-react';
 
 export default defineConfig({
   base: '/games',
   server: {
     port: 3001,
-    host: true
+    host: true,
   },
   plugins: [react() as PluginOption],
-})
+});