+++ apps/login-screen/CHANGELOG.md
@@ -1,5 +1,11 @@
 # login-screen
 
+## 0.1.3
+
+### Patch Changes
+
+- b1efb10: chore(42): update eslint for all apps
+
 ## 0.1.2
 
 ### Patch Changes